Fix: call isValidPassword when signing up and changing password
