+++ source/Login/Login.html
@@ -116,6 +116,7 @@
             </div>
         </div>
         <script src="../Backend/FirebaseInit.js" type="module"></script>
+        <script src="../Backend/BackendInit.js" type="module"></script>
         <script src="./Login.js" type="module"></script>
     </body>
 </html>
